Move more common logic into base
